pyproject.toml
[project]
name = "medicalagent"
version = "0.1.0"
description = ""
authors = [
    {name = "Mai Thi Huong",email = "hvgiang86@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12, <4.0"
dependencies = [
    "langchain (>=0.3.23,<0.4.0)",
    "sentence-transformers (>=4.1.0,<5.0.0)",
    "chromadb (>=1.0.6,<2.0.0)",
    "langchain-ollama (>=0.3.2,<0.4.0)",
    "langchain-community (>=0.3.21,<0.4.0)",
    "pypdf (>=5.4.0,<6.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "streamlit (>=1.29.0,<2.0.0)",
    "rich (>=13.7.0,<14.0.0)"
]

[tool.poetry]
package-mode = false


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
